Create initial header
I imported the font which will be used throughout the project,
now my configuration files also take care of my fonts. I created
a head function which renders the initial divs which will be used as
the tabs. they may change or the logic can change in a future commit,
maybe hard-coded into the html